refactor message handling to use a deterministic # of goroutines (not 3 per message) and use priority queues to manage the in flight and deferred requeue lists on a nsq channel

+++ nsqd/server_protocol_v2.go
@@ -282,7 +282,7 @@ func (p *ServerProtocolV2) REQ(client *nsq.ServerClient, params []string) ([]byt
 
 	channelInterface, _ := client.GetState("channel")
 	channel := channelInterface.(*Channel)
-	err = channel.RequeueMessage([]byte(idStr), timeoutMs)
+	err = channel.RequeueMessage([]byte(idStr), int64(timeoutMs))
 	if err != nil {
 		return nil, nsq.ClientErrV2RequeueFailed
 	}
